+++ circuit-foundation/src/commonMain/kotlin/com/slack/circuit/foundation/NavigableCircuitContent.kt
@@ -3,7 +3,6 @@
 package com.slack.circuit.foundation
 
 import androidx.compose.animation.AnimatedContent
-import androidx.compose.animation.AnimatedContentTransitionScope
 import androidx.compose.animation.ContentTransform
 import androidx.compose.animation.SizeTransform
 import androidx.compose.animation.core.CubicBezierEasing
@@ -19,13 +18,12 @@ import androidx.compose.animation.togetherWith
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.Immutable
+import androidx.compose.runtime.compositionLocalOf
 import androidx.compose.runtime.currentCompositeKeyHash
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.movableContentOf
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberUpdatedState
-import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import com.slack.circuit.backstack.BackStack
@@ -39,9 +37,9 @@ import com.slack.circuit.retained.LocalCanRetainChecker
 import com.slack.circuit.retained.LocalRetainedStateRegistry
 import com.slack.circuit.retained.RetainedStateRegistry
 import com.slack.circuit.retained.rememberRetained
+import com.slack.circuit.runtime.InternalCircuitApi
 import com.slack.circuit.runtime.Navigator
 import com.slack.circuit.runtime.screen.Screen
-import kotlin.math.sign
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.ImmutableMap
 import kotlinx.collections.immutable.toImmutableList
@@ -49,22 +47,22 @@ import kotlinx.collections.immutable.toImmutableList
 @Composable
 public fun NavigableCircuitContent(
   navigator: Navigator,
-  backstack: BackStack<out Record>,
+  backStack: BackStack<out Record>,
   modifier: Modifier = Modifier,
   circuit: Circuit = requireNotNull(LocalCircuit.current),
-  providedValues: ImmutableMap<out Record, ProvidedValues> = providedValuesForBackStack(backstack),
+  providedValues: ImmutableMap<out Record, ProvidedValues> = providedValuesForBackStack(backStack),
   decoration: NavDecoration = circuit.defaultNavDecoration,
   unavailableRoute: (@Composable (screen: Screen, modifier: Modifier) -> Unit) =
     circuit.onUnavailableContent,
 ) {
   val activeContentProviders =
-    backstack.buildCircuitContentProviders(
+    backStack.buildCircuitContentProviders(
       navigator = navigator,
       circuit = circuit,
       unavailableRoute = unavailableRoute,
     )
 
-  if (backstack.isEmpty) return
+  if (backStack.isEmpty) return
 
   /*
    * We store the RetainedStateRegistries for each back stack entry into an 'navigation content'
@@ -102,12 +100,12 @@ public fun NavigableCircuitContent(
   val outerRegistry = rememberRetained(key = outerKey) { RetainedStateRegistry() }
 
   CompositionLocalProvider(LocalRetainedStateRegistry provides outerRegistry) {
-    decoration.DecoratedContent(activeContentProviders, backstack.size, modifier) { provider ->
+    decoration.DecoratedContent(activeContentProviders, backStack.size, modifier) { provider ->
       // We retain the record's retained state registry for as long as the back stack
       // contains the record
       val record = provider.record
       val recordInBackStackRetainChecker =
-        remember(backstack, record) { CanRetainChecker { record in backstack } }
+        remember(backStack, record) { CanRetainChecker { record in backStack } }
 
       CompositionLocalProvider(LocalCanRetainChecker provides recordInBackStackRetainChecker) {
         // Remember the `providedValues` lookup because this composition can live longer than
@@ -123,6 +121,7 @@ public fun NavigableCircuitContent(
         CompositionLocalProvider(
           LocalRetainedStateRegistry provides recordRetainedStateRegistry,
           LocalCanRetainChecker provides CanRetainChecker.Always,
+          LocalBackStack provides backStack,
           *providedLocals,
         ) {
           provider.content(record)
@@ -221,9 +220,19 @@ public object NavigatorDefaults {
   // Mirrors the forward and backward transitions of activities in Android 34
   public object DefaultDecoration : NavDecoration {
 
-    private val forward: ContentTransform = computeTransition(1)
+    /**
+     * The [ContentTransform] used for 'forward' navigation changes (i.e. items added to stack).
+     * This isn't meant for public consumption, so be aware that this may be removed/changed at any
+     * time.
+     */
+    @InternalCircuitApi public val forward: ContentTransform by lazy { computeTransition(1) }
 
-    private val backward: ContentTransform = computeTransition(-1)
+    /**
+     * The [ContentTransform] used for 'backward' navigation changes (i.e. items popped off stack).
+     * This isn't meant for public consumption, so be aware that this may be removed/changed at any
+     * time.
+     */
+    @InternalCircuitApi public val backward: ContentTransform by lazy { computeTransition(-1) }
 
     private fun computeTransition(sign: Int): ContentTransform {
       val enterTransition =
@@ -262,18 +271,6 @@ public object NavigatorDefaults {
       return enterTransition togetherWith exitTransition
     }
 
-    private fun AnimatedContentTransitionScope<*>.transitionFor(diff: Int): ContentTransform {
-      return when {
-        diff > 0 -> forward
-        diff < 0 -> backward
-        else -> fadeIn() togetherWith fadeOut()
-      }.using(
-        // Disable clipping since the faded slide-in/out should
-        // be displayed out of bounds.
-        SizeTransform(clip = false)
-      )
-    }
-
     @Composable
     override fun <T> DecoratedContent(
       args: ImmutableList<T>,
@@ -281,14 +278,27 @@ public object NavigatorDefaults {
       modifier: Modifier,
       content: @Composable (T) -> Unit,
     ) {
-      // Remember the previous stack depth so we know if the navigation is going "back".
-      val prevStackDepth = rememberSaveable { mutableStateOf(backStackDepth) }
-      val diff = backStackDepth - prevStackDepth.value
-      prevStackDepth.value = backStackDepth
+      @OptIn(InternalCircuitApi::class)
       AnimatedContent(
         targetState = args,
         modifier = modifier,
-        transitionSpec = { transitionFor(diff) },
+        transitionSpec = {
+          // A transitionSpec should only use values passed into the `AnimatedContent`, to minimize
+          // the transitionSpec recomposing. The states are available as `targetState` and
+          // `initialState`
+          val diff = targetState.size - initialState.size
+          val sameRoot = targetState.lastOrNull() == initialState.lastOrNull()
+
+          when {
+            sameRoot && diff > 0 -> forward
+            sameRoot && diff < 0 -> backward
+            else -> fadeIn() togetherWith fadeOut()
+          }.using(
+            // Disable clipping since the faded slide-in/out should
+            // be displayed out of bounds.
+            SizeTransform(clip = false)
+          )
+        },
       ) {
         content(it.first())
       }
@@ -308,3 +318,9 @@ public object NavigatorDefaults {
     }
   }
 }
+
+/**
+ * Internal API to access the [BackStack] from within a [CircuitContent] or
+ * [rememberAnsweringNavigator] composable, useful for cases where we create nested nav handling.
+ */
+internal val LocalBackStack = compositionLocalOf<BackStack<out Record>?> { null }