+++ circuit-runtime-screen/src/jsMain/kotlin/com/slack/circuit/runtime/screen/PopResult.js.kt
@@ -1,0 +1,7 @@
+// Copyright (C) 2024 Slack Technologies, LLC
+// SPDX-License-Identifier: Apache-2.0
+package com.slack.circuit.runtime.screen
+
+import androidx.compose.runtime.Immutable
+
+@Immutable public actual interface PopResult