+++ circuitx/effects/src/commonMain/kotlin/com/slack/circuitx/effects/ImpressionEffect.kt
@@ -12,6 +12,7 @@ import androidx.compose.runtime.setValue
 import com.slack.circuit.retained.RetainedStateRegistry
 import com.slack.circuit.retained.rememberRetained
 import com.slack.circuit.runtime.Navigator
+import com.slack.circuit.runtime.screen.PopResult
 import com.slack.circuit.runtime.screen.Screen
 
 /**
@@ -70,23 +71,22 @@ public fun rememberImpressionNavigator(
 }
 
 private class OnNavEventNavigator(val delegate: Navigator, val onNavEvent: () -> Unit) : Navigator {
-
   override fun goTo(screen: Screen) {
     onNavEvent()
     delegate.goTo(screen)
   }
 
-  override fun pop(): Screen? {
+  override fun pop(result: PopResult?): Screen? {
     onNavEvent()
     return delegate.pop()
   }
 
-  override fun peek(): Screen? {
-    return delegate.peek()
-  }
+  override fun peek(): Screen? = delegate.peek()
 
-  override fun resetRoot(newRoot: Screen): List<Screen> {
+  override fun peekBackStack(): List<Screen> = delegate.peekBackStack()
+
+  override fun resetRoot(newRoot: Screen, saveState: Boolean, restoreState: Boolean): List<Screen> {
     onNavEvent()
-    return delegate.resetRoot(newRoot)
+    return delegate.resetRoot(newRoot, saveState, restoreState)
   }
 }